Resolve failing tests
